+++ wazo_calld/plugins/applications/notifier.py
@@ -11,6 +11,8 @@ from .schema import (
 )
 from .events import (
     CallAnswered,
+    CallContactingStarted,
+    CallContactingStopped,
     CallDeleted,
     CallEntered,
     CallInitiated,
@@ -64,6 +66,18 @@ class ApplicationNotifier:
         logger.debug('Application (%s): Call (%s) answered', application_uuid, call.id_)
         call = application_call_schema.dump(call).data
         event = CallAnswered(application_uuid, call)
+        self._bus.publish(event)
+
+    def call_contacting_started(self, application_uuid, call):
+        logger.debug('Application (%s): Call (%s) contacting started', application_uuid, call.id_)
+        call = application_call_schema.dump(call).data
+        event = CallContactingStarted(application_uuid, call)
+        self._bus.publish(event)
+
+    def call_contacting_stopped(self, application_uuid, call):
+        logger.debug('Application (%s): Call (%s) contacting stopped', application_uuid, call.id_)
+        call = application_call_schema.dump(call).data
+        event = CallContactingStopped(application_uuid, call)
         self._bus.publish(event)
 
     def destination_node_created(self, application_uuid, node):